requirements.txt
openai>=1.40.0
httpx>=0.27.0
python-dotenv>=1.0.0
rich>=13.0.0
pyyaml>=6.0.0
psycopg2-binary>=2.9.0
pgvector>=0.3.0
